# Optimize queue_status in AWS