--- a/cms/service/EvaluationService.py
+++ b/cms/service/EvaluationService.py
@@ -80,10 +80,9 @@ class EvaluationExecutor(Executor):
         # Lock used to guard the currently executing operations
         self._current_execution_lock = gevent.lock.RLock()
 
-
         # As evaluate operations are split by testcases, there are too
         # many entries in the queue to display, so we just take only one
-        # operation of each (type, object_id, dataset_id) tuple.
+        # operation of each (type, object_id, dataset_id, priority) tuple.
         # This dictionary maps any such tuple to a "queue entry" (lacking
         # the testcase codename) and keeps track of multiplicity.
         self.queue_status_cumulative = dict()
@@ -1027,7 +1026,9 @@ class EvaluationService(TriggeredService):
 
         As evaluate operations are split by testcases, there are too
         many entries in the queue to display, so we collect entries with the
-        same (type, object_id, dataset_id) tuple.
+        same (type, object_id, dataset_id, priority) tuple.
+        Generally, we will see only one evaluate operation for each submission
+        in the queue status.
 
         The entries are then ordered by priority and timestamp (the
         same criteria used to look at what to complete next).