Any instance of the AWS overview page in a browser calls the ``queue_status()`` RPC function of the evaluation service once every 5 seconds.

The function returns a list of entries in the queue, but collects evaluations of the same submission on the same dataset for different testcases:

> Evaluating the result of submission 1 on dataset 1 [16 time(s) in queue]

Currently, the collection process is performed from scratch every time ``queue_status`` is called by looping over all entries in the queue. The running time of ``queue_status()`` is therefore roughly proportional to the number of entries in the queue. This can be substantial when (re-)evaluating many submissions for tasks with many testcases. (In a test, we observed running times of about 0.5s for 100 submissions with 1000 test cases each. Note that since the function is called by every browser every 5 seconds, opening the AWS overview page on 10 tabs can therefore exhaust the evaluation service and slow down evaluation.)

We propose to not perform the collection process every time ``queue_status`` is called, but every time an item is added to or removed from the queue. The running time of ``queue_status`` is then roughly proportional to the number of entries in the collected table (the size of the output sent over the network). So the running time decreases by a factor of roughly the number of testcases.

For simplicity, we no longer keep track of the testcase numbers for each (submission,dataset) pair. Previously, the queue on AWS also listed one enqueued testcase for each such group. However, since queue entries were processed in heap order, the displayed testcase was only guaranteed to be the lowest-numbered one for the first queue entry. We didn't find displaying such a "random" testcase particularly useful. If desired, one could keep track of the lowest-numbered testcase for each (submission,dataset) pair by using a separate priority queue for each such pair (instead of only keeping track of the "multiplicity").

Remark: The function ``_missing_operations`` also appears to have running time essentially proportional to the size of the queue, but at least it is only called once every 117 seconds by ES.

**Acknowledgements:** The occasionally large running time was noticed by @t-lenz. The proposal was developed together with @magula.

<!-- Reviewable:start -->
---
This change is [<img src="https://reviewable.io/review_button.svg" height="34" align="absmiddle" alt="Reviewable"/>](https://reviewable.io/reviews/cms-dev/cms/1170)
<!-- Reviewable:end -->