@fagu 2021-01-07
We also no longer display the multiplicity for anything other than evaluations. (In all other cases, multiplicity should always by 1.)