@fagu 2021-01-16
Thanks for the review! Most things should be fixed. (See above.)