@andreyv 2021-01-25
Thanks! Everything looks good now. I tested it in different scenarios and it works fine.

I can't edit the branch, so I'll do this before the merge tomorrow:
1. Rename `remove_from_cumulative_status` to `_remove_from_cumulative_status` (private method)
2. Replace `str(priority)` with `priority` (not a reference, so `str` not needed)